(setEncoding, publish): Clarify Javadoc. Add comments to explain interaction with the inherited implementation.
(close): Report ErrorManager.CLOSE_FAILURE instead of FLUSH_FAILURE if
a closed stream is closed again.

(close): Emit head and tail strings even if no record has ever been
published to the stream.  The previous implementation caused XML log
files to be empty, which is different to the behavior of Sun J2SE 1.4.