Added try-catch-blocks to
 * Socket.java: socket()
 * ServerSocket.java: socket()
 * DatagramSocket.java: socket()

with clean-up of file-descriptors in case bind(), connect() or listen()
fail. This fix will avoid loss of file-descriptors in the case create()
is executed successfully, but bind/connect/listen fail with an exception.
Then close() was not called and the file-descriptor allcoated by create()
was not returned to the OS.