Fix building with --disable-tools
Building with --disable-tools is broken since commit [1]. After the
changes in this commit, the AC_PROG_JAVAC_WORKS macro in configure.ac
is not expanded, since it is defined as a "one-shot" macro (using
AC_DEFUN_ONCE), and used internally by AC_PROG_JAVA_WORKS, which is
expanded earlier. Thus if AC_PROG_JAVA_WORKS does not run (which is
the case with --disable-tools), JAVAC is never defined.

Additionally, these macros are not used as intended:

- AC_PROG_JAVA[C] are the ones that should be used. The *_WORKS macros
  are documented as being "for internal use only"
- AC_PROG_JAVAC must always be called before AC_PROG_JAVA

Fix this by reverting commit [1], and use the AC_PROG_JAVA[C] macros
as intended.

Additionally, make sure the GCJ_JAVA and ECJ_JAVA are always defined,
even if javac is not used.

 [1]: e6c6a0238291db27ccdaee229bda59d7981a3182