Make VMChannel honour the DSYNC flag
The native VMChannel open() method ignores the FileChannelImpl.DSYNC
mode flag. This flag is used when creating a RandomAccessFile instance
with the mode string set to "rwd". Because the flag is ignored, this
is equivalent to passing "rw" as the mode string, i.e. updates to the
file's content will not be written synchronously to the underlying
storage device.

Fix this by making the native VMChannel implementation honour the
DSYNC flag.

Fixes #3 (BZ#70659)