Process.destroy: Fix InternalError if native process not found
Process.destroy() may arbitrarily throw InternalError for "recently
finished" processes.

Process.destroy() ends up calling VMProcess.nativeKill(), which in
turn calls cpproc_kill(). If the process does not exist anymore,
cpproc_kill() returns ESRCH, which results in an InternalError being
thrown. But this can happen under normal conditions, if the process
is reaped just before the call to nativeKill(). This is obviously
not expected by user code.

Note that the Javadocs for Process.destroy() have been amended to
state that "If the process is not alive, no action is taken."

Fix this by handling ESRCH as success in VMProcess.nativeKill().

Fixes #6 (BZ#108569)