Process.destroy: <init>/destroy should not clobber interrupt flag
VMProcess.<init> and destroy() contain code to wait for a specific
process state. If the interrupt flag was already set in the calling
thread when this code was reached, or if the thread was interrupted
while waiting, the interrupt flag would be incorrectly cleared. Fix
this by explicitly restoring the interrupt flag on exit if needed.

Fixes #7 (BZ#109152)