VMProcess: Notify reaper thread in waitFor() and destroy()
The VMProcess class uses a dedicated helper thread that does all
the fork()'ing and wait()'ing for external processes. When waiting
for processes to finish, this thread sleeps for short periods (1 s),
and periodically polls for reapable children. This introduces delays
of up to 1 second in waitFor() and destroy().

A partial fix for this is to awake the helper thread from waitFor()
and destroy(). This avoids the unnecessary delay in cases where the
native processes have already finished when waitFor()/destroy() are
called.

Partially fixes #23 (BZ#111587)