Fix rebuild of vm/reference not being triggered
Trying to recompile after modifying vm/reference classes does not
correctly trigger a rebuild of those. This seems to be broken since
984a0dd6, which tried to fix 689e97df but missed the case where
vm_omitlist is empty, thus not remaking the AWK script and incorrectly
using the previous one for VM classes.

This was fixed by calling AWK when vm_omit is not empty, and otherwise
(when there are no classes exclussions) adding the whole list of VM
classes to the output file.