DateFormatSymbols: Fix parsing of strings with a trailing delimiter
Some strings in the locale data seem to contain an extraneous
trailing delimiter. This results in an incorrect value for the
last parsed symbol (the additional delimiter is returned as part
of the symbol).

Fix this by adjusting the 'limit' parameter that is passed to
Pattern.split(), so that any additional trailing data is not
included in the parsed symbols.

Fixes #5 (BZ#83968)